refactor: centralize artifact upload helpers
